ai-service/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.27.8
transformers==4.30.2
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
pillow>=9.0.0